requirements.txt
flask
flask-cors
torch
torchvision
ftfy
regex
timm
clip-by-openai
pillow
